initial commit: implement avalanche proof generation in python
This include functions from electrum and Electron Cash.
Everything is implemented in pure python, to make packaging easier.
libsecp256k1 can be used later when a proper python package is available
for it.

Test Plan:

in main.py, specify the private key for the utxo, and run this script.
The generated proof is checked against a proof build with Bitcoin ABC's
`bitcoin-cli buildavalancheproof ...` command.